+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2025-08-08
+### Added
+- Rotation field to images
+
+### Changed
+- Improved performance of image particles
+
 ## [1.2.3] - 2024-07-24
 ### Added
 - Particle permission to allow further configuration (allowed by default)